Add files via upload
data-request.py -- Bug fixed. When events have non ascii characters in the name, if the observations of the same event were distributed across several input files, the event was processed several times.

parameters-selection.py -- Bug fixed related to the the adjustment of the grid for searching parameters.

supervised.py -- Hipeparameters for boosting and random forest are not multiplied by two.